i3s: getTileColor as external function
